Move phases to stages
